Move Mergeable interface to SPI
