+++ core/trino-main/src/main/java/io/trino/operator/HashCollisionsInfo.java
@@ -15,7 +15,7 @@ package io.trino.operator;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.trino.util.Mergeable;
+import io.trino.spi.Mergeable;
 
 public class HashCollisionsInfo
         implements Mergeable<HashCollisionsInfo>, OperatorInfo
