+++ core/trino-main/src/main/java/io/trino/operator/join/JoinOperatorInfo.java
@@ -17,7 +17,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.trino.operator.OperatorInfo;
 import io.trino.operator.join.LookupJoinOperatorFactory.JoinType;
-import io.trino.util.Mergeable;
+import io.trino.spi.Mergeable;
 
 import java.util.Optional;
