+++ core/trino-spi/src/main/java/io/trino/spi/Mergeable.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.util;
+package io.trino.spi;
 
 public interface Mergeable<T>
 {
